requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
